% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{graph_perf}
\alias{graph_perf}
\title{Graphique de performances standard}
\usage{
graph_perf(
  data,
  obs,
  sim,
  .variable = NA,
  .poids = NULL,
  .effet_dev = FALSE,
  .couleur = "blue",
  .zoom = FALSE
)
}
\arguments{
\item{data}{\code{data.frame} des données en entrée ...}

\item{obs}{Colonne des observés.}

\item{sim}{Colonne des simulés.}

\item{.variable}{Nom de la variable étudié, pour nommer les axes. facultatif.}

\item{.poids}{Poids de chaque mesure. facultatif.}

\item{.effet_dev}{Y a t'il un effet développement de la variable étudiée ?
Utilisé pour le choix des indicateurs à afficher :
\itemize{
\item \code{FALSE} - RMSE, RRMSE, biais, N
\item \code{TRUE} - MAPE, N
}}

\item{.couleur}{couleur des points}

\item{.zoom}{tronquage des axes}
}
\value{
Graphique ggplot2 observé / simulé avec indicateurs
}
\description{
Graphique ggplot2 observé / simulé avec indicateurs de performances.
}
