% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{sauv_graph}
\alias{sauv_graph}
\title{Sauvegarde une figure en png et/ou svg}
\usage{
sauv_graph(graph, fichier, width, height, .res = 200, .silence = FALSE)
}
\arguments{
\item{graph}{figure}

\item{fichier}{emplacement ou enregistrer la figure
.png ou .svg à la fin pour limiter l'enregistrement à un seul format}

\item{width}{largeur en inches}

\item{height}{hauteur en inches}

\item{.res}{résolution pour le png, en ppi}

\item{.silence}{Désactive le témoin d'activation "."}
}
\description{
Sauvegarde une figure en png et/ou svg
}
